Use syscalls conditionally only for linux
